+++ rollup.config.js
@@ -1,11 +1,12 @@
 import {terser} from 'rollup-plugin-terser';
-import babel from 'rollup-plugin-babel';
+import babel from '@rollup/plugin-babel';
 
 function getConfig ({format = 'umd', min = false}) {
   return {
     input: 'src/index.js',
     plugins: [
       babel({
+        babelHelpers: 'bundled',
         plugins: ['transform-object-rest-spread']
       }),
       min ? terser() : undefined